+++ include/nstd/core/cstr.h
@@ -17,6 +17,12 @@ NSTDAPI NSTDSlice nstd_core_cstr_as_slice(const NSTDChar *const cstr);
 /// Returns: `NSTDUSize len` - The length of the C string.
 NSTDAPI NSTDUSize nstd_core_cstr_len(const NSTDChar *const cstr);
 
+/// Returns the length (in bytes) of a null terminated C string including the null terminator.
+/// Parameters:
+///     `const NSTDChar *const cstr` - The C string.
+/// Returns: `NSTDUSize len` - The length of the C string with the null terminator.
+NSTDAPI NSTDUSize nstd_core_cstr_len_with_null(const NSTDChar *const cstr);
+
 /// Compares two C strings and returns `NSTD_BOOL_TRUE` if they contain the same data.
 /// Parameters:
 ///     `const NSTDChar *cstr1` - The first C string.