+++ src/core/cstr.rs
@@ -29,6 +29,16 @@ pub unsafe extern "C" fn nstd_core_cstr_len(cstr: *const NSTDChar) -> usize {
     chr.offset_from(cstr) as usize
 }
 
+/// Returns the length (in bytes) of a null terminated C string including the null terminator.
+/// Parameters:
+///     `const NSTDChar *const cstr` - The C string.
+/// Returns: `NSTDUSize len` - The length of the C string with the null terminator.
+#[inline]
+#[cfg_attr(feature = "clib", no_mangle)]
+pub unsafe extern "C" fn nstd_core_cstr_len_with_null(cstr: *const NSTDChar) -> usize {
+    nstd_core_cstr_len(cstr) + 1
+}
+
 /// Compares two C strings and returns `NSTD_BOOL_TRUE` if they contain the same data.
 /// Parameters:
 ///     `const NSTDChar *cstr1` - The first C string.