+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Latest
+# 0.9.4
 ### `nstd.alloc`
 - `NSTDAllocator` now owns an `NSTDErrorCode`.
 ### `nstd.core`